refactor: abstract connectedChainId
