+++ src/components/Trade/hooks/useTradeRoutes/useTradeRoutes.ts
@@ -1,10 +1,9 @@
 import { Asset } from '@shapeshiftoss/asset-service'
-import { AssetId, ethChainId, fromAssetId, fromChainId } from '@shapeshiftoss/caip'
+import { AssetId, ethChainId, fromAssetId } from '@shapeshiftoss/caip'
 import { ChainAdapterManager } from '@shapeshiftoss/chain-adapters'
-import { ETHWallet } from '@shapeshiftoss/hdwallet-core'
 import { KnownChainIds } from '@shapeshiftoss/types'
 import isEmpty from 'lodash/isEmpty'
-import { useCallback, useEffect, useMemo, useState } from 'react'
+import { useCallback, useEffect } from 'react'
 import { useFormContext } from 'react-hook-form'
 import { useSelector } from 'react-redux'
 import { useHistory } from 'react-router-dom'
@@ -12,7 +11,6 @@ import { TradeAmountInputField, TradeRoutePaths, TradeState } from 'components/T
 import { getChainAdapters } from 'context/PluginProvider/PluginProvider'
 import { useEvm } from 'hooks/useEvm/useEvm'
 import { useWallet } from 'hooks/useWallet/useWallet'
-import { bnOrZero } from 'lib/bignumber/bignumber'
 import { selectAssetById, selectAssets } from 'state/slices/selectors'
 import { useAppSelector } from 'state/store'
 
@@ -38,21 +36,7 @@ export const useTradeRoutes = (
     state: { wallet },
   } = useWallet()
 
-  // fixme: abstract to custom hook
-  const [evmChainId, setEvmChainId] = useState<string | null>(null)
-  const { supportedEvmChainIds } = useEvm()
-
-  useEffect(() => {
-    ;(async () => {
-      const chainId = await (wallet as ETHWallet)?.ethGetChainId?.()
-      if (chainId) setEvmChainId(bnOrZero(chainId).toString())
-    })()
-  }, [wallet])
-
-  const connectedChainId = useMemo(
-    () => supportedEvmChainIds.find(chainId => fromChainId(chainId).chainReference === evmChainId),
-    [evmChainId, supportedEvmChainIds],
-  )
+  const { connectedChainId } = useEvm()
 
   const [defaultSellAssetId, defaultBuyAssetId] = getDefaultPair(connectedChainId)
