+++ src/hooks/useEvm/useEvm.ts
@@ -1,10 +1,15 @@
 import { CHAIN_NAMESPACE, fromChainId } from '@shapeshiftoss/caip'
-import { useMemo } from 'react'
+import { ETHWallet } from '@shapeshiftoss/hdwallet-core'
+import { bnOrZero } from '@shapeshiftoss/investor-foxy'
+import { useEffect, useMemo, useState } from 'react'
 import { getChainAdapters } from 'context/PluginProvider/PluginProvider'
+import { useWallet } from 'hooks/useWallet/useWallet'
 import { selectFeatureFlags } from 'state/slices/preferencesSlice/selectors'
 import { useAppSelector } from 'state/store'
 
 export const useEvm = () => {
+  const { state } = useWallet()
+  const [evmChainId, setEvmChainId] = useState<string | null>(null)
   const featureFlags = useAppSelector(selectFeatureFlags)
   const supportedEvmChainIds = useMemo(
     () =>
@@ -16,5 +21,17 @@ export const useEvm = () => {
     [featureFlags],
   )
 
-  return { supportedEvmChainIds }
+  useEffect(() => {
+    ;(async () => {
+      const chainId = await (state.wallet as ETHWallet)?.ethGetChainId?.()
+      if (chainId) setEvmChainId(bnOrZero(chainId).toString())
+    })()
+  }, [state])
+
+  const connectedChainId = useMemo(
+    () => supportedEvmChainIds.find(chainId => fromChainId(chainId).chainReference === evmChainId),
+    [evmChainId, supportedEvmChainIds],
+  )
+
+  return { supportedEvmChainIds, connectedChainId }
 }
